Default to mystery man icon when querying iconizer
Previously, when querying iconizer, if `icon.img` was undefined, the default mystery man icon in the item would get
replaced with `undefined`. This would break the logic for updating the icon with the D&D Beyond icons later down the
line when adding the NPCs to the compendium. The fix I made was simply to default to mystery-man when iconizer is used.